[update] 将common下的response和form分类成forms和responses
