+++ api/v1/hadoop/hadoop.go
@@ -1,7 +1,8 @@
 package hadoop
 
 import (
-	"Kube-CC/common"
+	"Kube-CC/common/forms"
+	"Kube-CC/common/responses"
 	"Kube-CC/service"
 	"github.com/gin-gonic/gin"
 	"go.uber.org/zap"
@@ -18,7 +19,7 @@ func Index(c *gin.Context) {
 	if u_id != "" {
 		uid, err = strconv.Atoi(u_id)
 		if err != nil {
-			c.JSON(http.StatusOK, common.Response{
+			c.JSON(http.StatusOK, responses.Response{
 				StatusCode: -1,
 				StatusMsg:  err.Error(),
 			})
@@ -27,7 +28,7 @@ func Index(c *gin.Context) {
 	}
 	hadoopListResponse, err := service.GetHadoop(uint(uid))
 	if err != nil {
-		c.JSON(http.StatusOK, common.Response{
+		c.JSON(http.StatusOK, responses.Response{
 			StatusCode: -1,
 			StatusMsg:  err.Error(),
 		})
@@ -39,15 +40,15 @@ func Index(c *gin.Context) {
 // Add 创建hadoop
 func Add(c *gin.Context) {
 	// 表单验证
-	form := common.HadoopAddForm{}
+	form := forms.HadoopAddForm{}
 	if err := c.ShouldBind(&form); err != nil {
-		c.JSON(http.StatusOK, common.ValidatorResponse(err))
+		c.JSON(http.StatusOK, responses.ValidatorResponse(err))
 		return
 	}
 
 	res, err := service.CreateHadoop(form.Uid, form.HdfsMasterReplicas, form.DatanodeReplicas, form.YarnMasterReplicas, form.YarnNodeReplicas, form.ExpiredTime, form.Cpu, form.Memory)
 	if err != nil {
-		c.JSON(http.StatusOK, common.Response{
+		c.JSON(http.StatusOK, responses.Response{
 			StatusCode: -1,
 			StatusMsg:  err.Error(),
 		})
@@ -61,7 +62,7 @@ func Delete(c *gin.Context) {
 	ns := c.Param("name")
 	res, err := service.DeleteHadoop(ns)
 	if err != nil {
-		c.JSON(http.StatusOK, common.Response{
+		c.JSON(http.StatusOK, responses.Response{
 			StatusCode: -1,
 			StatusMsg:  err.Error(),
 		})
@@ -72,9 +73,9 @@ func Delete(c *gin.Context) {
 
 func Update(c *gin.Context) {
 	// 表单验证
-	form := common.HadoopUpdateForm{}
+	form := forms.HadoopUpdateForm{}
 	if err := c.ShouldBind(&form); err != nil {
-		c.JSON(http.StatusOK, common.ValidatorResponse(err))
+		c.JSON(http.StatusOK, responses.ValidatorResponse(err))
 		return
 	}
 	uid := ""
@@ -83,7 +84,7 @@ func Update(c *gin.Context) {
 	}
 	res, err := service.UpdateHadoop(form.Name, uid, form.HdfsMasterReplicas, form.DatanodeReplicas, form.YarnMasterReplicas, form.YarnNodeReplicas, form.ExpiredTime, form.Cpu, form.Memory)
 	if err != nil {
-		c.JSON(http.StatusOK, common.Response{
+		c.JSON(http.StatusOK, responses.Response{
 			StatusCode: -1,
 			StatusMsg:  err.Error(),
 		})
@@ -95,9 +96,9 @@ func Update(c *gin.Context) {
 // BatchAdd 批量添加
 func BatchAdd(c *gin.Context) {
 	// 表单验证
-	form := common.BatchHadoopAddForm{}
+	form := forms.BatchHadoopAddForm{}
 	if err := c.ShouldBind(&form); err != nil {
-		c.JSON(http.StatusOK, common.ValidatorResponse(err))
+		c.JSON(http.StatusOK, responses.ValidatorResponse(err))
 		return
 	}
 	ids := form.Uid
@@ -112,5 +113,5 @@ func BatchAdd(c *gin.Context) {
 		}(id)
 	}
 	group.Wait()
-	c.JSON(http.StatusOK, common.OK)
+	c.JSON(http.StatusOK, responses.OK)
 }
