+++ api/v1/namespace/namespace.go
@@ -1,7 +1,8 @@
 package namespace
 
 import (
-	"Kube-CC/common"
+	"Kube-CC/common/forms"
+	"Kube-CC/common/responses"
 	"Kube-CC/service"
 	"github.com/gin-gonic/gin"
 	"go.uber.org/zap"
@@ -24,7 +25,7 @@ func Index(c *gin.Context) {
 
 	nsListResponse, err := service.GetNs(selector)
 	if err != nil {
-		c.JSON(http.StatusOK, common.Response{StatusCode: -1, StatusMsg: err.Error()})
+		c.JSON(http.StatusOK, responses.Response{StatusCode: -1, StatusMsg: err.Error()})
 	} else {
 		c.JSON(http.StatusOK, nsListResponse)
 	}
@@ -35,23 +36,23 @@ func Delete(c *gin.Context) {
 	ns := c.Param("ns")
 	response, err := service.DeleteNs(ns)
 	if err != nil {
-		c.JSON(http.StatusOK, common.Response{StatusCode: -1, StatusMsg: err.Error()})
+		c.JSON(http.StatusOK, responses.Response{StatusCode: -1, StatusMsg: err.Error()})
 	} else {
 		c.JSON(http.StatusOK, response)
 	}
 }
 
 func Add(c *gin.Context) {
-	form := common.NsAddForm{}
+	form := forms.NsAddForm{}
 	if err := c.ShouldBind(&form); err != nil {
-		c.JSON(http.StatusOK, common.ValidatorResponse(err))
+		c.JSON(http.StatusOK, responses.ValidatorResponse(err))
 		return
 	}
 	label := map[string]string{}
 	if form.Uid != 0 {
 		label["u_id"] = strconv.Itoa(int(form.Uid))
 	}
-	//expiredTime, err := time.Parse("2006-01-02 15:04:05", form.ExpiredTime)
+	//expiredTime, err := time.Parse("2006-01-02 15:04:05", forms.ExpiredTime)
 	//if err != nil {
 	//	zap.S().Errorln(err)
 	//	c.JSON(http.StatusOK, common.Response{StatusCode: -1, StatusMsg: err.Error()})
@@ -60,7 +61,7 @@ func Add(c *gin.Context) {
 	response, err := service.CreateNs(form.Name, form.ExpiredTime, label, form.Cpu, form.Memory, form.Num)
 	if err != nil {
 		zap.S().Errorln(err)
-		c.JSON(http.StatusOK, common.Response{StatusCode: -1, StatusMsg: err.Error()})
+		c.JSON(http.StatusOK, responses.Response{StatusCode: -1, StatusMsg: err.Error()})
 		return
 	}
 	c.JSON(http.StatusOK, response)
@@ -68,16 +69,16 @@ func Add(c *gin.Context) {
 
 // Update 更新namespace及其所含所有资源的uid
 func Update(c *gin.Context) {
-	form := common.NsAddForm{}
+	form := forms.NsAddForm{}
 	if err := c.ShouldBind(&form); err != nil {
-		c.JSON(http.StatusOK, common.ValidatorResponse(err))
+		c.JSON(http.StatusOK, responses.ValidatorResponse(err))
 		return
 	}
 	uid := ""
 	if form.Uid != 0 {
 		uid = strconv.Itoa(int(form.Uid))
 	}
-	//expiredTime, err := time.Parse("2006-01-02 15:04:05", form.ExpiredTime)
+	//expiredTime, err := time.Parse("2006-01-02 15:04:05", forms.ExpiredTime)
 	//if err != nil {
 	//	zap.S().Errorln(err)
 	//	c.JSON(http.StatusOK, common.Response{StatusCode: -1, StatusMsg: err.Error()})
@@ -86,7 +87,7 @@ func Update(c *gin.Context) {
 	response, err := service.UpdateNs(form.Name, uid, form.ExpiredTime, form.Cpu, form.Memory, form.Num)
 	if err != nil {
 		zap.S().Errorln(err)
-		c.JSON(http.StatusOK, common.Response{StatusCode: -1, StatusMsg: err.Error()})
+		c.JSON(http.StatusOK, responses.Response{StatusCode: -1, StatusMsg: err.Error()})
 	} else {
 		c.JSON(http.StatusOK, response)
 	}
