+++ api/v1/spark/spark.go
@@ -1,7 +1,8 @@
 package spark
 
 import (
-	"Kube-CC/common"
+	"Kube-CC/common/forms"
+	"Kube-CC/common/responses"
 	"Kube-CC/service"
 	"github.com/gin-gonic/gin"
 	"go.uber.org/zap"
@@ -18,7 +19,7 @@ func Index(c *gin.Context) {
 	if u_id != "" {
 		uid, err = strconv.Atoi(u_id)
 		if err != nil {
-			c.JSON(http.StatusOK, common.Response{
+			c.JSON(http.StatusOK, responses.Response{
 				StatusCode: -1,
 				StatusMsg:  err.Error(),
 			})
@@ -28,7 +29,7 @@ func Index(c *gin.Context) {
 
 	sparkListRes, err := service.GetSpark(uint(uid))
 	if err != nil {
-		c.JSON(http.StatusOK, common.Response{
+		c.JSON(http.StatusOK, responses.Response{
 			StatusCode: -1,
 			StatusMsg:  err.Error(),
 		})
@@ -40,14 +41,14 @@ func Index(c *gin.Context) {
 // Add 创建spark
 func Add(c *gin.Context) {
 	// 表单验证
-	form := common.SparkAddForm{}
+	form := forms.SparkAddForm{}
 	if err := c.ShouldBind(&form); err != nil {
-		c.JSON(http.StatusOK, common.ValidatorResponse(err))
+		c.JSON(http.StatusOK, responses.ValidatorResponse(err))
 		return
 	}
 	response, err := service.CreateSpark(form.Uid, form.MasterReplicas, form.WorkerReplicas, form.ExpiredTime, form.Cpu, form.Memory)
 	if err != nil {
-		c.JSON(http.StatusOK, common.Response{StatusCode: -1, StatusMsg: err.Error()})
+		c.JSON(http.StatusOK, responses.Response{StatusCode: -1, StatusMsg: err.Error()})
 		return
 	}
 	c.JSON(http.StatusOK, response)
@@ -57,7 +58,7 @@ func Delete(c *gin.Context) {
 	ns := c.Param("name")
 	response, err := service.DeleteSpark(ns)
 	if err != nil {
-		c.JSON(http.StatusOK, common.Response{StatusCode: -1, StatusMsg: err.Error()})
+		c.JSON(http.StatusOK, responses.Response{StatusCode: -1, StatusMsg: err.Error()})
 		return
 	}
 	c.JSON(http.StatusOK, response)
@@ -65,9 +66,9 @@ func Delete(c *gin.Context) {
 
 func Update(c *gin.Context) {
 	// 表单验证
-	form := common.SparkUpdateForm{}
+	form := forms.SparkUpdateForm{}
 	if err := c.ShouldBind(&form); err != nil {
-		c.JSON(http.StatusOK, common.ValidatorResponse(err))
+		c.JSON(http.StatusOK, responses.ValidatorResponse(err))
 		return
 	}
 	uid := ""
@@ -76,7 +77,7 @@ func Update(c *gin.Context) {
 	}
 	response, err := service.UpdateSpark(form.Name, uid, form.MasterReplicas, form.WorkerReplicas, form.ExpiredTime, form.Cpu, form.Memory)
 	if err != nil {
-		c.JSON(http.StatusOK, common.Response{StatusCode: -1, StatusMsg: err.Error()})
+		c.JSON(http.StatusOK, responses.Response{StatusCode: -1, StatusMsg: err.Error()})
 		return
 	}
 	c.JSON(http.StatusOK, response)
@@ -85,9 +86,9 @@ func Update(c *gin.Context) {
 // BatchAdd 批量添加
 func BatchAdd(c *gin.Context) {
 	// 表单验证
-	form := common.BatchSparkAddForm{}
+	form := forms.BatchSparkAddForm{}
 	if err := c.ShouldBind(&form); err != nil {
-		c.JSON(http.StatusOK, common.ValidatorResponse(err))
+		c.JSON(http.StatusOK, responses.ValidatorResponse(err))
 		return
 	}
 	ids := form.Uid
@@ -102,5 +103,5 @@ func BatchAdd(c *gin.Context) {
 		}(id)
 	}
 	group.Wait()
-	c.JSON(http.StatusOK, common.OK)
+	c.JSON(http.StatusOK, responses.OK)
 }
