+++ api/v1/svc/service.go
@@ -1,7 +1,7 @@
 package svc
 
 import (
-	"Kube-CC/common"
+	"Kube-CC/common/responses"
 	"Kube-CC/service"
 	"github.com/gin-gonic/gin"
 	"k8s.io/apimachinery/pkg/labels"
@@ -22,7 +22,7 @@ func Index(c *gin.Context) {
 	}
 	serviceListResponse, err := service.GetService(ns, selector)
 	if err != nil {
-		c.JSON(http.StatusOK, common.Response{StatusCode: -1, StatusMsg: err.Error()})
+		c.JSON(http.StatusOK, responses.Response{StatusCode: -1, StatusMsg: err.Error()})
 	} else {
 		c.JSON(http.StatusOK, serviceListResponse)
 	}
@@ -33,7 +33,7 @@ func Delete(c *gin.Context) {
 	name := c.Query("name")
 	response, err := service.DeleteService(name, ns)
 	if err != nil {
-		c.JSON(http.StatusOK, common.Response{StatusCode: -1, StatusMsg: err.Error()})
+		c.JSON(http.StatusOK, responses.Response{StatusCode: -1, StatusMsg: err.Error()})
 	} else {
 		c.JSON(http.StatusOK, response)
 	}
@@ -44,10 +44,10 @@ func Info(c *gin.Context) {
 	name := c.Query("name")
 	res, err := service.GetAService(name, ns)
 	if err != nil {
-		c.JSON(http.StatusOK, common.Response{StatusCode: -1, StatusMsg: err.Error()})
+		c.JSON(http.StatusOK, responses.Response{StatusCode: -1, StatusMsg: err.Error()})
 	} else {
-		c.JSON(http.StatusOK, common.ServiceInfo{
-			Response: common.OK,
+		c.JSON(http.StatusOK, responses.ServiceInfo{
+			Response: responses.OK,
 			Info:     *res,
 		})
 	}
