+++ api/v1/yaml/yaml.go
@@ -1,7 +1,8 @@
 package yaml
 
 import (
-	"Kube-CC/common"
+	"Kube-CC/common/forms"
+	"Kube-CC/common/responses"
 	"Kube-CC/service/yamlApply"
 	"encoding/json"
 	"errors"
@@ -15,9 +16,9 @@ func Apply(c *gin.Context) {
 	//body := c.Request.Body
 	//bytes, _ := ioutil.ReadAll(body)
 	//fmt.Println((bytes))
-	form := common.YamlApplyForm{}
+	form := forms.YamlApplyForm{}
 	if err := c.ShouldBindJSON(&form); err != nil {
-		c.JSON(http.StatusOK, common.ValidatorResponse(err))
+		c.JSON(http.StatusOK, responses.ValidatorResponse(err))
 		return
 	}
 	yaml := form.Yaml.(map[string]interface{})
@@ -82,24 +83,24 @@ func Apply(c *gin.Context) {
 		}
 		goto SUCCESS
 	default:
-		c.JSON(http.StatusOK, common.Response{
+		c.JSON(http.StatusOK, responses.Response{
 			StatusCode: -1,
 			StatusMsg:  "类型不匹配",
 		})
 		return
 	}
 SUCCESS:
-	c.JSON(http.StatusOK, common.OK)
+	c.JSON(http.StatusOK, responses.OK)
 	return
 END:
-	c.JSON(http.StatusOK, common.Response{StatusCode: -1, StatusMsg: err.Error()})
+	c.JSON(http.StatusOK, responses.Response{StatusCode: -1, StatusMsg: err.Error()})
 	return
 }
 
 func Create(c *gin.Context) {
-	form := common.YamlCreateForm{}
+	form := forms.YamlCreateForm{}
 	if err := c.ShouldBindJSON(&form); err != nil {
-		c.JSON(http.StatusOK, common.ValidatorResponse(err))
+		c.JSON(http.StatusOK, responses.ValidatorResponse(err))
 		return
 	}
 	yaml := form.Yaml.(map[string]interface{})
@@ -184,7 +185,7 @@ func Create(c *gin.Context) {
 		}
 		goto SUCCESS
 	default:
-		c.JSON(http.StatusOK, common.Response{
+		c.JSON(http.StatusOK, responses.Response{
 			StatusCode: -1,
 			StatusMsg:  "类型不匹配",
 		})
@@ -192,13 +193,13 @@ func Create(c *gin.Context) {
 	}
 
 SUCCESS:
-	c.JSON(http.StatusOK, common.OK)
+	c.JSON(http.StatusOK, responses.OK)
 	return
 END:
-	c.JSON(http.StatusOK, common.Response{StatusCode: -1, StatusMsg: err.Error()})
+	c.JSON(http.StatusOK, responses.Response{StatusCode: -1, StatusMsg: err.Error()})
 	return
 }
 
 func Delete(c *gin.Context) {
-	c.JSON(http.StatusOK, common.Response{1, "delete"})
+	c.JSON(http.StatusOK, responses.Response{1, "delete"})
 }
