+++ common/forms/deploy.go
@@ -1,0 +1,7 @@
+package forms
+
+type DeployAddForm struct {
+	Name      string `json:"name" forms:"name"`
+	Namespace string `json:"namespace" forms:"namespace"`
+	Uid       uint   `json:"u_id" forms:"u_id"`
+}
