+++ service/pod.go
@@ -1,7 +1,7 @@
 package service
 
 import (
-	"Kube-CC/common"
+	"Kube-CC/common/responses"
 	"Kube-CC/dao"
 	"context"
 	corev1 "k8s.io/api/core/v1"
@@ -18,17 +18,17 @@ func GetAPod(name, ns string) (*corev1.Pod, error) {
 }
 
 // GetPod 获得指定namespace下pod
-func GetPod(ns string, label string) (*common.PodListResponse, error) {
+func GetPod(ns string, label string) (*responses.PodListResponse, error) {
 	pods, err := dao.ClientSet.CoreV1().Pods(ns).List(context.Background(), metav1.ListOptions{LabelSelector: label})
 	if err != nil {
 		return nil, err
 	}
 
 	num := len(pods.Items)
-	podList := make([]common.Pod, num)
+	podList := make([]responses.Pod, num)
 
 	for i, pod := range pods.Items {
-		tmp := common.Pod{
+		tmp := responses.Pod{
 			Name:              pod.Name,
 			Namespace:         pod.Namespace,
 			CreatedAt:         pod.CreationTimestamp.Format("2006-01-02 15:04:05"),
@@ -39,16 +39,16 @@ func GetPod(ns string, label string) (*common.PodListResponse, error) {
 		}
 		podList[i] = tmp
 	}
-	return &common.PodListResponse{Response: common.OK, Length: num, PodList: podList}, nil
+	return &responses.PodListResponse{Response: responses.OK, Length: num, PodList: podList}, nil
 }
 
 // DeletePod 删除指定pod
-func DeletePod(name, ns string) (*common.Response, error) {
+func DeletePod(name, ns string) (*responses.Response, error) {
 	err := dao.ClientSet.CoreV1().Pods(ns).Delete(context.Background(), name, metav1.DeleteOptions{})
 	if err != nil {
 		return nil, err
 	}
-	return &common.OK, nil
+	return &responses.OK, nil
 }
 
 func CreatePod(name, ns string, label map[string]string, spec corev1.PodSpec) (*corev1.Pod, error) {
@@ -71,10 +71,10 @@ func CreatePod(name, ns string, label map[string]string, spec corev1.PodSpec) (*
 	return create, nil
 }
 
-func UpdatePod(pod *corev1.Pod) (*common.Response, error) {
+func UpdatePod(pod *corev1.Pod) (*responses.Response, error) {
 	_, err := dao.ClientSet.CoreV1().Pods(pod.Name).Update(context.Background(), pod, metav1.UpdateOptions{})
 	if err != nil {
 		return nil, err
 	}
-	return &common.OK, nil
+	return &responses.OK, nil
 }
