+++ service/yamlApply/apply.go
@@ -1,7 +1,7 @@
 package yamlApply
 
 import (
-	"Kube-CC/common"
+	"Kube-CC/common/responses"
 	"Kube-CC/dao"
 	"Kube-CC/service"
 	"context"
@@ -12,14 +12,14 @@ import (
 )
 
 // NamespaceApply namespace的更新或新建
-func NamespaceApply(ns *corev1.Namespace) (*common.Response, error) {
+func NamespaceApply(ns *corev1.Namespace) (*responses.Response, error) {
 	name := ns.Name
 	if _, err := dao.ClientSet.CoreV1().Namespaces().Get(context.Background(), name, metav1.GetOptions{}); err != nil {
 		if errors.IsNotFound(err) {
 			if _, err := dao.ClientSet.CoreV1().Namespaces().Create(context.Background(), ns, metav1.CreateOptions{}); err != nil {
 				return nil, err
 			}
-			return &common.OK, nil
+			return &responses.OK, nil
 		} else {
 			return nil, err
 		}
@@ -27,13 +27,13 @@ func NamespaceApply(ns *corev1.Namespace) (*common.Response, error) {
 		if _, err := dao.ClientSet.CoreV1().Namespaces().Update(context.Background(), ns, metav1.UpdateOptions{}); err != nil {
 			return nil, err
 		}
-		return &common.OK, nil
+		return &responses.OK, nil
 	}
 
 }
 
 // DeployApply  deploy的更新或者创建
-func DeployApply(deploy *appsv1.Deployment) (*common.Response, error) {
+func DeployApply(deploy *appsv1.Deployment) (*responses.Response, error) {
 	name := deploy.Name
 	ns := deploy.Namespace
 	labels := deploy.Labels
@@ -56,7 +56,7 @@ func DeployApply(deploy *appsv1.Deployment) (*common.Response, error) {
 			if _, err := service.CreateDeploy(name, ns, labels, deploy.Spec); err != nil {
 				return nil, err
 			}
-			return &common.OK, nil
+			return &responses.OK, nil
 		} else { // 其他错误直接返回
 			return nil, err
 		}
@@ -68,7 +68,7 @@ func DeployApply(deploy *appsv1.Deployment) (*common.Response, error) {
 }
 
 // ServiceApply service的更新或者创建
-func ServiceApply(svc *corev1.Service) (*common.Response, error) {
+func ServiceApply(svc *corev1.Service) (*responses.Response, error) {
 	name := svc.Name
 	ns := svc.Namespace
 	labels := svc.Labels
@@ -90,7 +90,7 @@ func ServiceApply(svc *corev1.Service) (*common.Response, error) {
 			if _, err := service.CreateService(name, ns, labels, svc.Spec); err != nil {
 				return nil, err
 			}
-			return &common.OK, nil
+			return &responses.OK, nil
 		} else { // 其他错误直接返回
 			return nil, err
 		}
@@ -100,7 +100,7 @@ func ServiceApply(svc *corev1.Service) (*common.Response, error) {
 	}
 }
 
-func PodApply(pod *corev1.Pod) (*common.Response, error) {
+func PodApply(pod *corev1.Pod) (*responses.Response, error) {
 	name := pod.Name
 	ns := pod.Namespace
 	labels := pod.Labels
@@ -122,7 +122,7 @@ func PodApply(pod *corev1.Pod) (*common.Response, error) {
 			if _, err := service.CreatePod(name, ns, labels, pod.Spec); err != nil {
 				return nil, err
 			}
-			return &common.OK, nil
+			return &responses.OK, nil
 		} else { // 其他错误直接返回
 			return nil, err
 		}
